fix: Parse refs from bytes, not from String.
The latter can cause issues around illformed UTF-8 which wouldn't
bother git either.

This comes at the expense of not parsing line by line anymore, but
instead reading as fast as possible and parsing afterwards.

Performance wise I think it doesn't matter, but it will cause
more memory to be used. If this ever becomes a problem,
for example during pushes where we are stuck with V1, we can consider
implementing our own streaming appreach that works with packet lines
instead - they are just not exposed here even though they could.
