+++ git-protocol/Cargo.toml
@@ -46,7 +46,7 @@ git-credentials = { version = "^0.7.0", path = "../git-credentials" }
 
 thiserror = "1.0.32"
 serde = { version = "1.0.114", optional = true, default-features = false, features = ["derive"]}
-bstr = { version = "1.0.1", default-features = false, features = ["std"] }
+bstr = { version = "1.0.1", default-features = false, features = ["std", "unicode"] }
 nom = { version = "7", default-features = false, features = ["std"]}
 btoi = "0.4.2"
