--- git-protocol/src/fetch/tests/mod.rs
@@ -1,4 +1,0 @@
-#[cfg(any(feature = "async-client", feature = "blocking-client"))]
-mod arguments;
-#[cfg(any(feature = "blocking-client", feature = "async-client"))]
-mod refs;
