+++ git-protocol/src/handshake/refs/async_io.rs
@@ -1,19 +1,17 @@
 use futures_io::AsyncBufRead;
-use futures_lite::AsyncBufReadExt;
+use futures_lite::AsyncReadExt;
 
 use crate::handshake::{refs, refs::parse::Error, Ref};
+use bstr::ByteSlice;
 
 /// Parse refs from the given input line by line. Protocol V2 is required for this to succeed.
 pub async fn from_v2_refs(in_refs: &mut (dyn AsyncBufRead + Unpin)) -> Result<Vec<Ref>, Error> {
     let mut out_refs = Vec::new();
-    let mut line = String::new();
-    loop {
-        line.clear();
-        let bytes_read = in_refs.read_line(&mut line).await?;
-        if bytes_read == 0 {
-            break;
-        }
-        out_refs.push(refs::shared::parse_v2(&line)?);
+    let mut buf = Vec::new();
+
+    in_refs.read_to_end(&mut buf).await?;
+    for line in ByteSlice::lines(buf.as_slice()) {
+        out_refs.push(refs::shared::parse_v2(line.into())?);
     }
     Ok(out_refs)
 }
@@ -32,14 +30,11 @@ pub async fn from_v1_refs_received_as_part_of_handshake_and_capabilities<'a>(
 ) -> Result<Vec<Ref>, refs::parse::Error> {
     let mut out_refs = refs::shared::from_capabilities(capabilities)?;
     let number_of_possible_symbolic_refs_for_lookup = out_refs.len();
-    let mut line = String::new();
-    loop {
-        line.clear();
-        let bytes_read = in_refs.read_line(&mut line).await?;
-        if bytes_read == 0 {
-            break;
-        }
-        refs::shared::parse_v1(number_of_possible_symbolic_refs_for_lookup, &mut out_refs, &line)?;
+
+    let mut buf = Vec::new();
+    in_refs.read_to_end(&mut buf).await?;
+    for line in buf.as_slice().lines() {
+        refs::shared::parse_v1(number_of_possible_symbolic_refs_for_lookup, &mut out_refs, line.into())?;
     }
     Ok(out_refs.into_iter().map(Into::into).collect())
 }
