+++ git-protocol/src/handshake/refs/tests.rs
@@ -15,6 +15,8 @@ unborn refs/heads/symbolic symref-target:refs/heads/target
 "
     .as_bytes();
 
+    #[cfg(feature = "blocking-client")]
+    let input = &mut Fixture(input);
     let out = refs::from_v2_refs(input).await.expect("no failure on valid input");
 
     assert_eq!(
@@ -56,6 +58,7 @@ unborn refs/heads/symbolic symref-target:refs/heads/target
 
 #[maybe_async::test(feature = "blocking-client", async(feature = "async-client", async_std::test))]
 async fn extract_references_from_v1_refs() {
+    #[cfg_attr(feature = "blocking-client", allow(unused_mut))]
     let input = &mut "73a6868963993a3328e7d8fe94e5a6ac5078a944 HEAD
 21c9b7500cb144b3169a6537961ec2b9e865be81 MISSING_NAMESPACE_TARGET
 73a6868963993a3328e7d8fe94e5a6ac5078a944 refs/heads/main
@@ -63,6 +66,8 @@ async fn extract_references_from_v1_refs() {
 dce0ea858eef7ff61ad345cc5cdac62203fb3c10 refs/tags/git-commitgraph-v0.0.0
 21c9b7500cb144b3169a6537961ec2b9e865be81 refs/tags/git-commitgraph-v0.0.0^{}"
         .as_bytes();
+    #[cfg(feature = "blocking-client")]
+    let input = &mut Fixture(input);
     let out = refs::from_v1_refs_received_as_part_of_handshake_and_capabilities(
         input,
         Capabilities::from_bytes(b"\0symref=HEAD:refs/heads/main symref=MISSING_NAMESPACE_TARGET:(null)")
@@ -106,7 +111,7 @@ fn extract_symbolic_references_from_capabilities() -> Result<(), client::Error> 
     let caps = client::Capabilities::from_bytes(
         b"\0unrelated symref=HEAD:refs/heads/main symref=ANOTHER:refs/heads/foo symref=MISSING_NAMESPACE_TARGET:(null) agent=git/2.28.0",
     )?
-    .0;
+        .0;
     let out = refs::shared::from_capabilities(caps.iter()).expect("a working example");
 
     assert_eq!(
@@ -127,4 +132,39 @@ fn extract_symbolic_references_from_capabilities() -> Result<(), client::Error> 
         ]
     );
     Ok(())
+}
+
+#[cfg(feature = "blocking-client")]
+struct Fixture<'a>(&'a [u8]);
+
+#[cfg(feature = "blocking-client")]
+impl<'a> std::io::Read for Fixture<'a> {
+    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
+        self.0.read(buf)
+    }
+}
+
+#[cfg(feature = "blocking-client")]
+impl<'a> std::io::BufRead for Fixture<'a> {
+    fn fill_buf(&mut self) -> std::io::Result<&[u8]> {
+        self.0.fill_buf()
+    }
+
+    fn consume(&mut self, amt: usize) {
+        self.0.consume(amt)
+    }
+}
+
+#[cfg(feature = "blocking-client")]
+impl<'a> git_transport::client::ReadlineBufRead for Fixture<'a> {
+    fn readline(
+        &mut self,
+    ) -> Option<std::io::Result<Result<git_packetline::PacketLineRef<'_>, git_packetline::decode::Error>>> {
+        use bstr::{BStr, ByteSlice};
+        let bytes: &BStr = self.0.into();
+        let mut lines = bytes.lines();
+        let res = lines.next()?;
+        self.0 = lines.as_bytes();
+        Some(Ok(Ok(git_packetline::PacketLineRef::Data(res))))
+    }
 }
